requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.5.0
black                         22.1.0
click                         8.0.4
Django                        4.0.3
django-cors-headers           3.11.0
djangorestframework           3.13.1
djangorestframework-simplejwt 5.1.0
flake8                        4.0.1
mccabe                        0.6.1
mypy-extensions               0.4.3
pathspec                      0.9.0
pip                           21.2.4
platformdirs                  2.5.1
psycopg2                      2.9.3
pycodestyle                   2.8.0
pyflakes                      2.4.0
pytz                          2021.3
setuptools                    58.1.0
sqlparse                      0.4.2
tomli                         2.0.1
